refactor: ViewController에서 BookInfoView 분리 및 적용
- 기존 book header UI를 BookInfoView로 분리하여 역할 명확화
- ViewController에서 book 관련 UI 요소 제거
- ViewController는 BookInfoView에 데이터만 전달하도록 개선
- BookInfoView에서 이미지, 타이틀, 저자, 출판일, 페이지 수 표시 처리
